Move block-related endpoints to block_handler and file rename
